Move CreatureLanguages into QuteCreature, and fix up generated documentation.
Add support for `MarkdownDoclet` to correctly format and parse record
components. This was actually a pain to figure out, as it turns out that
these aren't accessible as an Element and instead have to be parsed through
the @param tags of the class comment.
