+++ src/main/java/dev/ebullient/convert/tools/pf2e/qute/QuteDataArmorClass.java
@@ -1,11 +1,11 @@
 package dev.ebullient.convert.tools.pf2e.qute;
 
-import java.util.Collection;
-import java.util.stream.Collectors;
-
 import dev.ebullient.convert.qute.NamedText;
 import dev.ebullient.convert.qute.QuteUtil;
 import io.quarkus.qute.TemplateData;
+
+import java.util.Collection;
+import java.util.stream.Collectors;
 
 /**
  * Pf2eTools armor class attributes
@@ -18,7 +18,11 @@ public class QuteDataArmorClass implements QuteUtil {
     /** Notes associated with the armor class, e.g. "with mage armor". */
     public Collection<String> notes;
 
-    /** Any notes associated with the armor class. This contains the same data as {@link #notes}, but as a single semicolon-delimited string. */
+    /**
+     * Any notes associated with the armor class. This contains the same data as
+     * {@link dev.ebullient.convert.tools.pf2e.qute.QuteDataArmorClass#notes notes}, but as a single
+     * semicolon-delimited string.
+     */
     public String note;
 
     public String toString() {
